+++ R/plume.R
@@ -132,8 +132,11 @@ Plume <- R6Class(
     #' @return A character vector.
     get_author_list = function(suffix = NULL, format = deprecated()) {
       if (lifecycle::is_present(format)) {
-        lifecycle::deprecate_warn("0.2.0", "get_author_list(format)", "get_author_list(suffix)")
-        suffix <- format
+        lifecycle::deprecate_stop(
+          "0.2.1",
+          "get_author_list(format)",
+          "get_author_list(suffix)"
+        )
       }
       authors <- private$pull("literal_name")
       if (is_empty(suffix)) {
@@ -260,7 +263,7 @@ Plume <- R6Class(
         sep = sep,
         last = sep_last
       ), .by = all_of(pars$grp_var))
-      if ((are_credit_roles(private$roles) || private$crt) && !by_author) {
+      if (are_credit_roles(private$roles) && !by_author) {
         out <- arrange(out, role)
       }
       out <- collapse_cols(out, pars$format, sep = divider)