+++ tests/testthat/test-get-contributions.R
@@ -179,29 +179,6 @@ test_that("set_main_contributors() ranks contributors", {
   )
 })
 
-# Deprecation ----
-
-test_that("specifying roles inside columns is deprecated", {
-  expect_snapshot({
-    aut <- Plume$new(data.frame(
-      given_name = "Zip",
-      family_name = "Zap",
-      role = "a"
-    ))
-  })
-  expect_equal(aut$get_contributions(), "a: Z.Z.")
-})
-
-test_that("`credit_roles = TRUE` is deprecated", {
-  expect_snapshot({
-    aut <- Plume$new(
-      data.frame(given_name = "Zip", family_name = "Zap", analysis = 1),
-      credit_roles = TRUE
-    )
-  })
-  expect_equal(aut$get_contributions(), "Formal analysis: Z.Z.")
-})
-
 # Errors ----
 
 test_that("get_contributions() gives meaningful error messages", {