+++ tests/testthat/test-set-status.R
@@ -14,12 +14,6 @@ test_that("sets status to selected authors", {
   }, rep(TRUE, 3))
 
   expect_equal({
-    withr::local_options(lifecycle_verbosity = "quiet")
-    aut$set_corresponding_authors(everyone_but(ric), .by = "given_name")
-    aut$get_plume()$corresponding
-  }, c(TRUE, FALSE, TRUE))
-
-  expect_equal({
     aut$set_corresponding_authors(1)
     aut$get_plume()$corresponding
   }, c(TRUE, FALSE, FALSE))
@@ -47,24 +41,6 @@ test_that("sets status to selected authors", {
     aut$set_deceased(zip, .by = "given_name")
     aut$get_plume()$deceased
   }, c(TRUE, FALSE, FALSE))
-})
-
-# Deprecation ----
-
-test_that("the `by` parameter is deprecated", {
-  expect_snapshot({
-    aut <- Plume$new(basic_df)
-    aut$set_corresponding_authors(zip, by = "given_name")
-  })
-  expect_equal(aut$get_plume()$corresponding, c(TRUE, FALSE, FALSE))
-})
-
-test_that("set_equal_contributor() is deprecated", {
-  expect_snapshot({
-    aut <- PlumeQuarto$new(basic_df, temp_file())
-    aut$set_equal_contributor(1, 3)
-  })
-  expect_equal(aut$get_plume()$equal_contributor, c(TRUE, FALSE, TRUE))
 })
 
 # Errors ----
@@ -110,8 +86,5 @@ test_that("set_*() methods give meaningful error messages", {
 })
 
 test_that("everyone*() selectors error if used in a wrong context", {
-  expect_snapshot({
-    (expect_error(everyone()))
-    (expect_error(everyone_but()))
-  })
+  expect_snapshot(everyone(), error = TRUE)
 })